+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Add Custom Claims support for user JWT.
+- Add possibility to get of OAuth2 token by credentials.
+- Add `user_set_claims`.
+
 
 ## [0.6] - 2020-01-22